+++ temporalio/test/client_test.rb
@@ -140,30 +140,6 @@ class ClientTest < Test
     end
   end
 
-  class EmptyFirstHistoryEventsInterceptor
-    include Temporalio::Client::Interceptor
-
-    def intercept_client(next_interceptor)
-      Outbound.new(next_interceptor)
-    end
-
-    class Outbound < Temporalio::Client::Interceptor::Outbound
-      def initialize(next_interceptor)
-        super
-        @first_history_fetch = true
-      end
-
-      def fetch_workflow_history_events(input)
-        if @first_history_fetch
-          @first_history_fetch = false
-          return Enumerator.new { |_yielder| nil }
-        end
-
-        super
-      end
-    end
-  end
-
   def test_interceptor
     # Create client with interceptor
     track = TrackCallsInterceptor.new
@@ -213,20 +189,36 @@ class ClientTest < Test
     end
   end
 
-  def test_result_retries_when_history_fetch_returns_no_events
-    client = Temporalio::Client.new(
-      **env.client.options.with(interceptors: [EmptyFirstHistoryEventsInterceptor.new]).to_h
-    ) # steep:ignore
+  def test_result_retries_when_history_long_poll_returns_no_events
+    workflow_service = env.client.workflow_service
+    original_fetch = workflow_service.method(:get_workflow_execution_history)
+    empty_responses_remaining = 2
+    history_fetch_count = 0
+    workflow_service.define_singleton_method(:get_workflow_execution_history) do |request, **kwargs|
+      history_fetch_count += 1
+      if empty_responses_remaining.positive?
+        empty_responses_remaining -= 1
+        Temporalio::Api::WorkflowService::V1::GetWorkflowExecutionHistoryResponse.new
+      else
+        original_fetch.call(request, **kwargs)
+      end
+    end
     task_queue = "tq-#{SecureRandom.uuid}"
 
-    Temporalio::Worker.new(client:, task_queue:, workflows: [SimpleWorkflow]).run do
-      handle = client.start_workflow(
-        SimpleWorkflow,
-        'Temporal',
-        id: "wf-#{SecureRandom.uuid}",
-        task_queue:
-      )
-      assert_equal 'Hello, Temporal!', handle.result
+    begin
+      Temporalio::Worker.new(client: env.client, task_queue:, workflows: [SimpleWorkflow]).run do
+        handle = env.client.start_workflow(
+          SimpleWorkflow,
+          'Temporal',
+          id: "wf-#{SecureRandom.uuid}",
+          task_queue:
+        )
+        assert_equal 'Hello, Temporal!', handle.result
+      end
+      assert_equal 0, empty_responses_remaining
+      assert_operator history_fetch_count, :>=, 3
+    ensure
+      workflow_service.singleton_class.send(:remove_method, :get_workflow_execution_history)
     end
   end